Validator moved to Common as it is schema independent.
